cloud/requirements.txt
# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
aws-cdk-lib>=2.27.0, <3.0.0
cdk-aws-iotfleetwise>=0.2.12
constructs>=10.0.5, <11.0.0
